+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.0.0...HEAD)
 * [PCM-1795](https://inindca.atlassian.net/browse/PCM-1795) – Do not request media when `sdk.updateDefaultDevices({ updateActiveSessions: true, ... })`
   is called and the session is already using the requested deviceId.
+* [PCM-1798](https://inindca.atlassian.net/browse/PCM-1798) – Added logic to video session to check sure we received an initial conversation
+  event from hawk. This fixes the issue where sometimes joining a video session happens before we subscribe and receive the initial conversation
+    event from hawk – which leaves an empty list of participants in the session.
 
 # [v7.0.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v6.1.7...v7.0.0)
 ### BREAKING CHANGE